refactored, restructured and extended public api
