+++ gtest/stdout_print.cc
@@ -18,7 +18,7 @@ class stdout_print : public ::testing::Test
  protected:
   void TearDown() override
   {
-    auto& args = cuke::get_program_args(0, {});
+    auto& args = cuke::internal::get_program_args(0, {});
     args.clear();
     cuke::log::enable();
   }
@@ -194,10 +194,10 @@ TEST_F(stdout_print, scenario_from_file)
       std::format("{}/test_files/example.feature:3", unittests::test_dir());
   const char* argv[] = {"program", file_arg.c_str()};
   int argc = sizeof(argv) / sizeof(argv[0]);
-  cuke::program_args prog_args;
+  cuke::internal::program_args prog_args;
   prog_args.initialize(argc, argv);
 
-  const cuke::feature_file& file = prog_args.get_feature_files().back();
+  const cuke::internal::feature_file& file = prog_args.get_feature_files().back();
 
   cuke::parser p;
   p.parse_from_file(file);
@@ -337,9 +337,9 @@ TEST_F(stdout_print, scenario_fail_final_form_file)
   const char* argv[] = {"program", file_arg.c_str()};
   int argc = sizeof(argv) / sizeof(argv[0]);
 
-  cuke::program_args& prog_args = cuke::get_program_args(argc, argv);
+  cuke::internal::program_args& prog_args = cuke::internal::get_program_args(argc, argv);
 
-  const cuke::feature_file& file = prog_args.get_feature_files().back();
+  const cuke::internal::feature_file& file = prog_args.get_feature_files().back();
 
   cuke::parser p;
   p.parse_from_file(file);
@@ -365,8 +365,8 @@ TEST_F(stdout_print, scenario_fail_final_form_file_quiet)
   const char* argv[] = {"program", file_arg.c_str(), "--quiet"};
   int argc = sizeof(argv) / sizeof(argv[0]);
 
-  cuke::program_args& prog_args = cuke::get_program_args(argc, argv);
-  const cuke::feature_file& file = prog_args.get_feature_files().back();
+  cuke::internal::program_args& prog_args = cuke::internal::get_program_args(argc, argv);
+  const cuke::internal::feature_file& file = prog_args.get_feature_files().back();
 
   cuke::parser p;
   p.parse_from_file(file);
@@ -392,8 +392,8 @@ TEST_F(stdout_print, scenario_fail_final_form_file_q)
   const char* argv[] = {"program", file_arg.c_str(), "-q"};
   int argc = sizeof(argv) / sizeof(argv[0]);
 
-  cuke::program_args& prog_args = cuke::get_program_args(argc, argv);
-  const cuke::feature_file& file = prog_args.get_feature_files().back();
+  cuke::internal::program_args& prog_args = cuke::internal::get_program_args(argc, argv);
+  const cuke::internal::feature_file& file = prog_args.get_feature_files().back();
 
   cuke::parser p;
   p.parse_from_file(file);
@@ -419,7 +419,7 @@ TEST_F(stdout_print, verbose_tags)
 
   const char* argv[] = {"program", "-v", "-t", "@tag2"};
   int argc = sizeof(argv) / sizeof(argv[0]);
-  [[maybe_unused]] auto& args = cuke::get_program_args(argc, argv);
+  [[maybe_unused]] auto& args = cuke::internal::get_program_args(argc, argv);
 
   cuke::parser p;
   p.parse_script(script);
@@ -458,7 +458,7 @@ TEST_F(stdout_print, log_disabled_w_json_output)
 
   const char* argv[] = {"program", "--report-json"};
   int argc = sizeof(argv) / sizeof(argv[0]);
-  [[maybe_unused]] auto& args = cuke::get_program_args(argc, argv);
+  [[maybe_unused]] auto& args = cuke::internal::get_program_args(argc, argv);
 
   cuke::parser p;
   p.parse_script(script);
