+++ src/asserts.hpp
@@ -13,8 +13,8 @@ inline void internal_assert(bool condition, std::string_view error_msg)
 {
   if (condition == false)
   {
-    cuke::results::set_step_to(cuke::results::test_status::failed);
-    cuke::results::steps_back().error_msg = error_msg;
+    cuke::results::internal::set_step_to(cuke::results::test_status::failed);
+    cuke::results::internal::steps_back().error_msg = error_msg;
     cuke::log::info(log::color::red());
     cuke::log::info(error_msg);
     cuke::log::info(log::color::reset());
