+++ src/catalog.cpp
@@ -101,14 +101,15 @@ std::string as_json(std::size_t indents /* = 2 */)
 void print_readable_text_to_sink()
 {
   cuke::internal::write_to_file_or_stdout(
-      as_readable_text(), get_program_args().get_value(
-                              cuke::program_args::arg::steps_catalog_readable));
+      as_readable_text(),
+      cuke::internal::get_program_args().get_value(
+          cuke::internal::program_args::arg::steps_catalog_readable));
 }
 void print_json_to_sink(std::size_t indents /* = 2 */)
 {
   cuke::internal::write_to_file_or_stdout(
-      as_json(), get_program_args().get_value(
-                     cuke::program_args::arg::steps_catalog_json));
+      as_json(), cuke::internal::get_program_args().get_value(
+                     cuke::internal::program_args::arg::steps_catalog_json));
 }
 
 }  // namespace cuke::catalog
