+++ src/cucumber.cpp
@@ -56,7 +56,7 @@ cuke::results::test_status entry_point(int argc, const char* argv[])
 
 cwt_cucumber::cwt_cucumber(int argc, const char* argv[])
 {
-  get_program_args().initialize(argc, argv);
+  internal::get_program_args().initialize(argc, argv);
 }
 void cwt_cucumber::run_tests() const noexcept
 {
@@ -67,7 +67,8 @@ void cwt_cucumber::run_tests() const noexcept
 }
 void cwt_cucumber::print_results() const noexcept
 {
-  if (get_program_args().is_set(program_args::arg::report_json))
+  if (internal::get_program_args().is_set(
+          internal::program_args::arg::report_json))
   {
     report::print_json_to_sink();
   }
@@ -82,9 +83,9 @@ void cwt_cucumber::print_results() const noexcept
 
 bool cwt_cucumber::print_help() const noexcept
 {
-  if (get_program_args().is_set(program_args::arg::help))
+  if (internal::get_program_args().is_set(internal::program_args::arg::help))
   {
-    print_help_screen();
+    internal::print_help_screen();
     return true;
   }
   return false;
@@ -92,12 +93,14 @@ bool cwt_cucumber::print_help() const noexcept
 bool cwt_cucumber::export_catalog(
     std::size_t json_indents /* = 2 */) const noexcept
 {
-  if (get_program_args().is_set(program_args::arg::steps_catalog_readable))
+  if (internal::get_program_args().is_set(
+          internal::program_args::arg::steps_catalog_readable))
   {
     catalog::print_readable_text_to_sink();
     return true;
   }
-  if (get_program_args().is_set(program_args::arg::steps_catalog_json))
+  if (internal::get_program_args().is_set(
+          internal::program_args::arg::steps_catalog_json))
   {
     catalog::print_json_to_sink(json_indents);
     return true;
